+++ app/v1/models/dialation.py
@@ -1,9 +1,20 @@
 # -*- coding: utf-8 -*-
 import mongoengine as me
 from datetime import datetime
-from typing import Any, Dict
-from flask_restx import fields
+from flask_restx import fields, Namespace
 
+
+api = Namespace('dialation', description='Image dialation')
+model = api.model(
+    'Dialation',
+    {
+        'id': fields.String(required=True, description="dialation id"),
+        'iterations': fields.Integer(required=True, min=1, description="dialation iterations"),
+        'kernel_width': fields.Integer(required=True, min=1, description="kernel width"),
+        'kernel_height': fields.Integer(required=True, min=1, description="kernel height"),
+        'created_at': fields.String(required=True, description="created ISO datetime"),
+    }
+)
 class Dialation(me.Document):
 
     iterations = me.IntField(required=True, min_value=1)
@@ -19,13 +30,3 @@ class Dialation(me.Document):
         if not self.created_at:
             self.created_at = datetime.now().isoformat()
         return super().save(*args, **kwargs)
-
-    @staticmethod
-    def api_model() -> Dict[str, Any]:
-        return {
-            'id': fields.String(required=True, description="dialation id"),
-            'iterations': fields.Integer(required=True, min=1, description="dialation iterations"),
-            'kernel_width': fields.Integer(required=True, min=1, description="kernel width"),
-            'kernel_height': fields.Integer(required=True, min=1, description="kernel height"),
-            'created_at': fields.String(required=True, description="created ISO datetime"),
-        }
