--- app/v1/namespaces/apex.py
@@ -1,7 +1,0 @@
-from flask_restx import Namespace
-from app.v1.models.apex import Apex
-
-
-# define namespaces
-api = Namespace('apex', description='apex')
-apex_model = api.model('Apex', Apex.api_model())
