+++ app/v1/routes/captures.py
@@ -9,19 +9,19 @@ Copyright © 2021 LEAP. All Rights Reserved.
 
 import os
 import uuid
+
 from flask import current_app as app
 from flask import jsonify, make_response
 from flask_restx import Resource
 from mongoengine.errors import ValidationError
 from werkzeug.datastructures import FileStorage
-from app.v1.models.capture import Capture
-from app.v1.namespaces.capture import api, capture_model, post_model
+from app.v1.models.capture import Capture, api, model, post_model
 
 
 @api.route('')
 class CapturesList(Resource):
 
-    @api.marshal_with(capture_model, as_list=True, code=200)
+    @api.marshal_with(model, as_list=True, code=200)
     def get(self):
         """
         Get list of captures
@@ -29,7 +29,7 @@ class CapturesList(Resource):
         captures_list = Capture.objects()  # type: ignore
         return list(captures_list)
 
-    @api.marshal_with(capture_model, code=201)
+    @api.marshal_with(model, code=201)
     @api.expect(post_model, validate=True)
     def post(self):
         """
@@ -57,7 +57,7 @@ class CapturesList(Resource):
 @api.route('/<string:id>')
 class Captures(Resource):
 
-    @api.marshal_with(capture_model, code=200)
+    @api.marshal_with(model, code=200)
     def get(self, id: str):
         """
         Get a single capture
