--- app/v1/routes/users.py
@@ -1,10 +1,0 @@
-from flask_restful import Resource
-
-
-class Users(Resource):
-    def get(self, id: int): ...
-
-    def put(self, id: int): ...
-
-    def post(self): ...
-    def delete(self, id: int): ...
